# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  department: string;
  avatarInitials: string;
  color: string;
}

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  /** @minLength 1 */
  department: string;
  /** @nullable */
  color?: string | null;
}

export interface Meeting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: string;
  endTime: string;
  organizerId: number;
  /** @nullable */
  location?: string | null;
  isOnline?: boolean;
  /** @nullable */
  meetingLink?: string | null;
  createdAt: string;
  organizer: Employee;
  participants: Employee[];
}

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  startTime: string;
  endTime: string;
  organizerId: number;
  participantIds: number[];
  location?: string;
  isOnline?: boolean;
  meetingLink?: string;
}

export interface ConflictCheckInput {
  startTime: string;
  endTime: string;
  participantIds: number[];
  /** @nullable */
  excludeMeetingId?: number | null;
}

export type ConflictCheckResultConflictsItem = {
  employeeId: number;
  employeeName: string;
  conflictingMeetingId: number;
  conflictingMeetingTitle: string;
};

export interface ConflictCheckResult {
  hasConflict: boolean;
  conflicts: ConflictCheckResultConflictsItem[];
}

export type ConflictErrorConflictsItem = {
  employeeId: number;
  employeeName: string;
  conflictingMeetingId: number;
  conflictingMeetingTitle: string;
};

export interface ConflictError {
  error: string;
  conflicts: ConflictErrorConflictsItem[];
}

export interface Comment {
  id: number;
  meetingId: number;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  authorName: string;
  /** @minLength 1 */
  content: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetEmployeeMeetingsParams = {
from?: string;
to?: string;
};

export type ListMeetingsParams = {
from?: string;
to?: string;
/**
 * @nullable
 */
employeeId?: number | null;
search?: string;
/**
 * @nullable
 */
organizerId?: number | null;
status?: ListMeetingsStatus;
/**
 * @nullable
 */
isOnline?: boolean | null;
};

export type ListMeetingsStatus = typeof ListMeetingsStatus[keyof typeof ListMeetingsStatus];


export const ListMeetingsStatus = {
  upcoming: 'upcoming',
  completed: 'completed',
  all: 'all',
} as const;