# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Meeting calendar API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: employees
  - name: meetings
  - name: comments
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /employees/{id}/meetings:
    get:
      operationId: getEmployeeMeetings
      tags: [employees]
      summary: Get upcoming meetings for an employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meetings with optional filters
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: employeeId
          schema:
            type: ["integer", "null"]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: organizerId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: string
            enum: [upcoming, completed, all]
        - in: query
          name: isOnline
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: createMeeting
      tags: [meetings]
      summary: Create meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"

  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeeting
      tags: [meetings]
      summary: Delete meeting
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /meetings/today:
    get:
      operationId: getTodayMeetings
      tags: [meetings]
      summary: Today's meetings
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings/week:
    get:
      operationId: getWeekMeetings
      tags: [meetings]
      summary: This week's meetings
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"

  /meetings/check-conflict:
    post:
      operationId: checkConflict
      tags: [meetings]
      summary: Check scheduling conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConflictCheckInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictCheckResult"

  /meetings/{id}/comments:
    get:
      operationId: getMeetingComments
      tags: [comments]
      summary: Get comments for a meeting
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a meeting
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Employee:
      type: object
      required: [id, name, role, department, avatarInitials, color]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        department:
          type: string
        avatarInitials:
          type: string
        color:
          type: string

    EmployeeInput:
      type: object
      required: [name, role, department]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        department:
          type: string
          minLength: 1
        color:
          type: ["string", "null"]

    Meeting:
      type: object
      required: [id, title, startTime, endTime, organizerId, organizer, participants, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        organizerId:
          type: integer
        location:
          type: ["string", "null"]
        isOnline:
          type: boolean
        meetingLink:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        organizer:
          $ref: "#/components/schemas/Employee"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Employee"

    MeetingInput:
      type: object
      required: [title, startTime, endTime, organizerId, participantIds]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        organizerId:
          type: integer
        participantIds:
          type: array
          items:
            type: integer
        location:
          type: string
        isOnline:
          type: boolean
        meetingLink:
          type: string

    ConflictCheckInput:
      type: object
      required: [startTime, endTime, participantIds]
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        participantIds:
          type: array
          items:
            type: integer
        excludeMeetingId:
          type: ["integer", "null"]

    ConflictCheckResult:
      type: object
      required: [hasConflict, conflicts]
      properties:
        hasConflict:
          type: boolean
        conflicts:
          type: array
          items:
            type: object
            required: [employeeId, employeeName, conflictingMeetingId, conflictingMeetingTitle]
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              conflictingMeetingId:
                type: integer
              conflictingMeetingTitle:
                type: string

    ConflictError:
      type: object
      required: [error, conflicts]
      properties:
        error:
          type: string
        conflicts:
          type: array
          items:
            type: object
            required: [employeeId, employeeName, conflictingMeetingId, conflictingMeetingTitle]
            properties:
              employeeId:
                type: integer
              employeeName:
                type: string
              conflictingMeetingId:
                type: integer
              conflictingMeetingTitle:
                type: string

    Comment:
      type: object
      required: [id, meetingId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        meetingId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    CommentInput:
      type: object
      required: [authorName, content]
      properties:
        authorName:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string