# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */

export interface Comment {
  id: number;
  meetingId: number;
  authorName: string;
  content: string;
  createdAt: Date;
}