# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */

export interface CommentInput {
  /** @minLength 1 */
  authorName: string;
  /** @minLength 1 */
  content: string;
}