# lib/api-zod/src/generated/types/conflictCheckInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */

export interface ConflictCheckInput {
  startTime: Date;
  endTime: Date;
  participantIds: number[];
  /** @nullable */
  excludeMeetingId?: number | null;
}