# lib/api-zod/src/generated/types/conflictError.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */
import type { ConflictErrorConflictsItem } from './conflictErrorConflictsItem';

export interface ConflictError {
  error: string;
  conflicts: ConflictErrorConflictsItem[];
}