# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */
import type { Employee } from './employee';

export interface Meeting {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startTime: Date;
  endTime: Date;
  organizerId: number;
  /** @nullable */
  location?: string | null;
  isOnline?: boolean;
  /** @nullable */
  meetingLink?: string | null;
  createdAt: Date;
  organizer: Employee;
  participants: Employee[];
}