# lib/api-zod/src/generated/types/meetingInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Meeting calendar API
 * OpenAPI spec version: 0.2.0
 */

export interface MeetingInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  startTime: Date;
  endTime: Date;
  organizerId: number;
  participantIds: number[];
  location?: string;
  isOnline?: boolean;
  meetingLink?: string;
}